+++ components/camel-aws/camel-aws2-eventbridge/src/test/java/org/apache/camel/component/aws2/eventbridge/localstack/EventbridgeListRuleNamesByTargetIT.java
@@ -27,13 +27,12 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.eventbridge.EventbridgeConstants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.eventbridge.model.ListRulesResponse;
-import software.amazon.awssdk.services.eventbridge.model.RuleState;
+import software.amazon.awssdk.services.eventbridge.model.ListRuleNamesByTargetResponse;
 import software.amazon.awssdk.services.eventbridge.model.Target;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.Assert.assertEquals;
 
-public class EventbridgeEnableRuleLocalstackTest extends Aws2EventbridgeBaseTest {
+public class EventbridgeListRuleNamesByTargetIT extends Aws2EventbridgeBase {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -66,36 +65,20 @@ public class EventbridgeEnableRuleLocalstackTest extends Aws2EventbridgeBaseTest
             }
         });
 
-        template.send("direct:evs-disableRule", new Processor() {
+        template.send("direct:list-rule-name", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(EventbridgeConstants.RULE_NAME, "firstrule");
+                exchange.getIn().setHeader(EventbridgeConstants.TARGET_ARN,
+                        "arn:aws:sqs:eu-west-1:780410022472:camel-connector-test");
             }
         });
 
-        template.send("direct:evs-enableRule", new Processor() {
-
-            @Override
-            public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(EventbridgeConstants.RULE_NAME, "firstrule");
-            }
-        });
-
-        Exchange ex = template.request("direct:evs-listRules", new Processor() {
-
-            @Override
-            public void process(Exchange exchange) throws Exception {
-            }
-        });
-
-        ListRulesResponse resp = ex.getIn().getBody(ListRulesResponse.class);
-        assertEquals(true, resp.hasRules());
-        assertEquals(1, resp.rules().size());
-        assertEquals("firstrule", resp.rules().get(0).name());
-        assertEquals(RuleState.ENABLED, resp.rules().get(0).state());
         assertMockEndpointsSatisfied();
-
+        assertEquals(1, result.getExchanges().size());
+        assertEquals(1, result.getExchanges().get(0).getIn().getBody(ListRuleNamesByTargetResponse.class).ruleNames().size());
+        assertEquals("firstrule",
+                result.getExchanges().get(0).getIn().getBody(ListRuleNamesByTargetResponse.class).ruleNames().get(0));
     }
 
     @Override
@@ -103,18 +86,13 @@ public class EventbridgeEnableRuleLocalstackTest extends Aws2EventbridgeBaseTest
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                String putRule
+                String awsEndpoint
                         = "aws2-eventbridge://default?operation=putRule&eventPatternFile=file:src/test/resources/eventpattern.json";
-                String putTargets = "aws2-eventbridge://default?operation=putTargets";
-                String listRule = "aws2-eventbridge://default?operation=listRules";
-                String disableRule = "aws2-eventbridge://default?operation=disableRule";
-                String enableRule = "aws2-eventbridge://default?operation=enableRule";
-
-                from("direct:evs").to(putRule);
-                from("direct:evs-targets").to(putTargets);
-                from("direct:evs-listRules").to(listRule);
-                from("direct:evs-disableRule").to(disableRule);
-                from("direct:evs-enableRule").to(enableRule).log("${body}").to("mock:result");
+                String target = "aws2-eventbridge://default?operation=putTargets";
+                String listRule = "aws2-eventbridge://default?operation=listRuleNamesByTarget";
+                from("direct:evs").to(awsEndpoint);
+                from("direct:evs-targets").to(target);
+                from("direct:list-rule-name").to(listRule).to("mock:result");
             }
         };
     }
