+++ components/camel-aws/camel-aws2-eventbridge/src/test/java/org/apache/camel/component/aws2/eventbridge/localstack/EventbridgeListRulesIT.java
@@ -27,12 +27,12 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.eventbridge.EventbridgeConstants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.eventbridge.model.ListTargetsByRuleResponse;
+import software.amazon.awssdk.services.eventbridge.model.ListRulesResponse;
 import software.amazon.awssdk.services.eventbridge.model.Target;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class EventbridgeListTargetsByRuleLocalstackTest extends Aws2EventbridgeBaseTest {
+public class EventbridgeListRulesIT extends Aws2EventbridgeBase {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -65,18 +65,19 @@ public class EventbridgeListTargetsByRuleLocalstackTest extends Aws2EventbridgeB
             }
         });
 
-        template.send("direct:evs-list-targets", new Processor() {
+        Exchange ex = template.request("direct:evs-listRules", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(EventbridgeConstants.RULE_NAME, "firstrule");
             }
         });
+
+        ListRulesResponse resp = ex.getIn().getBody(ListRulesResponse.class);
+        assertEquals(true, resp.hasRules());
+        assertEquals(1, resp.rules().size());
+        assertEquals("firstrule", resp.rules().get(0).name());
         assertMockEndpointsSatisfied();
-        assertEquals(1, result.getExchanges().size());
-        assertEquals(1, result.getExchanges().get(0).getIn().getBody(ListTargetsByRuleResponse.class).targets().size());
-        assertEquals("sqs-queue",
-                result.getExchanges().get(0).getIn().getBody(ListTargetsByRuleResponse.class).targets().get(0).id());
+
     }
 
     @Override
@@ -84,13 +85,14 @@ public class EventbridgeListTargetsByRuleLocalstackTest extends Aws2EventbridgeB
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                String awsEndpoint
+                String putRule
                         = "aws2-eventbridge://default?operation=putRule&eventPatternFile=file:src/test/resources/eventpattern.json";
-                String target = "aws2-eventbridge://default?operation=putTargets";
-                String listTargets = "aws2-eventbridge://default?operation=listTargetsByRule";
-                from("direct:evs").to(awsEndpoint);
-                from("direct:evs-targets").to(target);
-                from("direct:evs-list-targets").to(listTargets).to("mock:result");
+                String putTargets = "aws2-eventbridge://default?operation=putTargets";
+                String listRule = "aws2-eventbridge://default?operation=listRules";
+
+                from("direct:evs").to(putRule).log("${body}");
+                from("direct:evs-targets").to(putTargets).log("${body}");
+                from("direct:evs-listRules").to(listRule).log("${body}").to("mock:result");
             }
         };
     }
