+++ components/camel-aws/camel-aws2-eventbridge/src/test/java/org/apache/camel/component/aws2/eventbridge/localstack/EventbridgeRemoveTargetsIT.java
@@ -27,13 +27,11 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.eventbridge.EventbridgeConstants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.eventbridge.model.ListRulesResponse;
-import software.amazon.awssdk.services.eventbridge.model.RuleState;
 import software.amazon.awssdk.services.eventbridge.model.Target;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.Assert.assertEquals;
 
-public class EventbridgeDisableRuleLocalstackTest extends Aws2EventbridgeBaseTest {
+public class EventbridgeRemoveTargetsIT extends Aws2EventbridgeBase {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -66,28 +64,18 @@ public class EventbridgeDisableRuleLocalstackTest extends Aws2EventbridgeBaseTes
             }
         });
 
-        template.send("direct:evs-disableRule", new Processor() {
+        template.send("direct:evs-remove-targets", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
                 exchange.getIn().setHeader(EventbridgeConstants.RULE_NAME, "firstrule");
+                List<String> targets = new ArrayList<String>();
+                targets.add("sqs-queue");
+                exchange.getIn().setHeader(EventbridgeConstants.TARGETS_IDS, targets);
             }
         });
-
-        Exchange ex = template.request("direct:evs-listRules", new Processor() {
-
-            @Override
-            public void process(Exchange exchange) throws Exception {
-            }
-        });
-
-        ListRulesResponse resp = ex.getIn().getBody(ListRulesResponse.class);
-        assertEquals(true, resp.hasRules());
-        assertEquals(1, resp.rules().size());
-        assertEquals("firstrule", resp.rules().get(0).name());
-        assertEquals(RuleState.DISABLED, resp.rules().get(0).state());
         assertMockEndpointsSatisfied();
-
+        assertEquals(1, result.getExchanges().size());
     }
 
     @Override
@@ -95,16 +83,13 @@ public class EventbridgeDisableRuleLocalstackTest extends Aws2EventbridgeBaseTes
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                String putRule
+                String awsEndpoint
                         = "aws2-eventbridge://default?operation=putRule&eventPatternFile=file:src/test/resources/eventpattern.json";
-                String putTargets = "aws2-eventbridge://default?operation=putTargets";
-                String listRule = "aws2-eventbridge://default?operation=listRules";
-                String disableRule = "aws2-eventbridge://default?operation=disableRule";
-
-                from("direct:evs").to(putRule);
-                from("direct:evs-targets").to(putTargets);
-                from("direct:evs-listRules").to(listRule);
-                from("direct:evs-disableRule").to(disableRule).log("${body}").to("mock:result");
+                String target = "aws2-eventbridge://default?operation=putTargets";
+                String removeTargets = "aws2-eventbridge://default?operation=removeTargets";
+                from("direct:evs").to(awsEndpoint);
+                from("direct:evs-targets").to(target);
+                from("direct:evs-remove-targets").to(removeTargets).to("mock:result");
             }
         };
     }
